Implement vacuum() methods for all table engines
- Added vacuum() to MemoryBTree: iterates BTreeMap and vacuums version chains
- Added vacuum() to PagedBTree: recursively vacuums tree nodes and writes back
- Added vacuum() to MemoryHashTable: iterates HashMap and vacuums version chains
- Added vacuum() to MemoryART: recursively vacuums ART nodes (all node types)
- Added vacuum() to LsmTree: vacuums active and immutable memtables
- Added vacuum() to MemoryGraphTable: vacuums storage + in-memory index (outgoing/incoming edges)
- Fixed TimeSeriesTable vacuum() return type from u64 to usize for consistency

All implementations:
- Take min_visible_lsn as parameter
- Call VersionChain::vacuum() on each entry
- Return total count of removed versions as usize
- Preserve one old version as base for efficiency

Closes nanokv-cod

+++ src/table/timeseries/mod.rs
@@ -285,13 +285,13 @@ impl<FS: FileSystem> TimeSeriesTable<FS> {
     }
 
     /// Vacuum old versions that are no longer visible.
-    pub fn vacuum(&self, min_visible_lsn: LogSequenceNumber) -> TableResult<u64> {
+    pub fn vacuum(&self, min_visible_lsn: LogSequenceNumber) -> TableResult<usize> {
         let mut state = self.state.write().unwrap();
-        let mut total_removed = 0u64;
+        let mut total_removed = 0usize;
 
         for manager in state.series.values_mut() {
             for bucket in manager.buckets.values_mut() {
-                total_removed += bucket.vacuum(min_visible_lsn) as u64;
+                total_removed += bucket.vacuum(min_visible_lsn);
             }
         }
